[squash] bf45040e: denc: add with_dencv to denc_traits
